+++ src/main/backend.ts
@@ -9,7 +9,7 @@ import {
   Scope,
 } from '@sentry/core';
 import { NodeBackend } from '@sentry/node';
-import { Event, EventHint, Severity, Transport, TransportOptions } from '@sentry/types';
+import { Event, EventHint, ScopeContext, Severity, Transport, TransportOptions } from '@sentry/types';
 import { Dsn, forget, logger, SentryError } from '@sentry/utils';
 import { app, BrowserWindow, crashReporter, ipcMain } from 'electron';
 import { join } from 'path';
@@ -29,7 +29,7 @@ function getCachePath(): string {
 }
 
 /**
- * Retruns a promise that resolves when app is ready.
+ * Returns a promise that resolves when app is ready.
  */
 export async function isAppReady(): Promise<boolean> {
   return (
@@ -40,6 +40,28 @@ export async function isAppReady(): Promise<boolean> {
       });
     })
   );
+}
+
+/** Is object defined and has keys */
+function hasKeys(obj: any): boolean {
+  return obj != undefined && Object.keys(obj).length > 0;
+}
+
+/** Gets a Scope object with user, tags and extra */
+function getScope(): Partial<ScopeContext> {
+  const scope = getCurrentHub().getScope() as any | undefined;
+
+  if (!scope) {
+    return {};
+  }
+
+  return {
+    /* eslint-disable @typescript-eslint/no-unsafe-member-access */
+    ...(hasKeys(scope._user) && { user: scope._user }),
+    ...(hasKeys(scope._tags) && { tags: scope._tags }),
+    ...(hasKeys(scope._extra) && { extra: scope._extra }),
+    /* eslint-enable @typescript-eslint/no-unsafe-member-access */
+  };
 }
 
 /** Backend implementation for Electron renderer backends. */
@@ -253,6 +275,11 @@ export class MainBackend extends BaseBackend<ElectronOptions> implements CommonB
       throw new SentryError('Attempted to enable Electron native crash reporter but no DSN was supplied');
     }
 
+    // We don't add globalExtra for Linux because Breakpad doesn't support JSON like strings:
+    // https://github.com/electron/electron/issues/29711
+    const globalExtra =
+      process.platform !== 'linux' ? { sentry___initialScope: JSON.stringify(getScope()) } : undefined;
+
     const dsn = new Dsn(dsnString);
 
     // We will manually submit errors, but CrashReporter requires a submitURL in
@@ -265,6 +292,7 @@ export class MainBackend extends BaseBackend<ElectronOptions> implements CommonB
       submitURL: MinidumpUploader.minidumpUrlFromDsn(dsn),
       uploadToServer: this._options.useCrashpadMinidumpUploader || false,
       compress: true,
+      globalExtra,
     });
 
     if (this._options.useSentryMinidumpUploader !== false) {
@@ -369,25 +397,21 @@ export class MainBackend extends BaseBackend<ElectronOptions> implements CommonB
       const sentScope = Scope.clone(rendererScope) as any;
       /* eslint-disable @typescript-eslint/no-unsafe-member-access */
       configureScope(scope => {
-        const isDefinedAndHasKeys = (obj: any): boolean => {
-          return obj != undefined && Object.keys(obj).length > 0;
-        };
-
-        if (isDefinedAndHasKeys(sentScope._user)) {
+        if (hasKeys(sentScope._user)) {
           scope.setUser(sentScope._user);
         }
 
-        if (isDefinedAndHasKeys(sentScope._tags)) {
+        if (hasKeys(sentScope._tags)) {
           scope.setTags(sentScope._tags);
         }
 
-        if (isDefinedAndHasKeys(sentScope._extra)) {
+        if (hasKeys(sentScope._extra)) {
           scope.setExtras(sentScope._extra);
         }
 
-        // Since we do not have updates for individual breadcrumbs anymore and only for the whole scope
+        // Since we do not have updates for individual breadcrumbs any more and only for the whole scope
         // we just add the last added breadcrumb on scope updates
-        scope.addBreadcrumb(sentScope._breadcrumbs.pop());
+        scope.addBreadcrumb(sentScope._breadcrumbs.pop(), this._options.maxBreadcrumbs);
       });
       /* eslint-enable @typescript-eslint/no-unsafe-member-access */
     });